requirements.txt
PyYAML~=5.3.1
pytest==6.2.4